feat: start rolling out seoContent in article web tests
