+++ packages/article/__tests__/images.native.js
@@ -1,3 +1,5 @@
+import React from "react";
+import TestRenderer from "react-test-renderer";
 import {
   addSerializers,
   compose,
@@ -5,8 +7,12 @@ import {
   minimaliseTransform,
   minimalNativeTransform
 } from "@times-components/jest-serializer";
+import { iterator } from "@times-components/test-utils";
+import Article from "../src/article";
+import articleFixture, { testFixture } from "../fixtures/full-article";
+import { adConfig } from "./ad-mock";
+import articleProps from "./shared-article-props";
 import "./mocks.native";
-import shared from "./images.base";
 
 const omitKeys = new Set([
   "data",
@@ -19,6 +25,16 @@ const omitKeys = new Set([
   "viewabilityConfig",
   "viewabilityConfigCallbackPairs"
 ]);
+
+const emptyArticle = {
+  byline: null,
+  flags: null,
+  label: null,
+  leadAsset: null,
+  relatedArticleSlice: null,
+  standfirst: null,
+  topics: null
+};
 
 export default () => {
   addSerializers(
@@ -44,5 +60,86 @@ export default () => {
     global.Intl = realIntl;
   });
 
-  shared();
+  const tests = [
+    {
+      name: "a secondary image",
+      test() {
+        const testInstance = TestRenderer.create(
+          <Article
+            {...articleProps}
+            adConfig={adConfig}
+            analyticsStream={() => {}}
+            article={articleFixture({
+              ...testFixture,
+              ...emptyArticle,
+              content: [
+                {
+                  attributes: {
+                    caption: "A Caption",
+                    credits: "Some Credits",
+                    display: "secondary",
+                    ratio: "3:2",
+                    url: "https://image-2.io"
+                  },
+                  children: [],
+                  name: "image"
+                }
+              ]
+            })}
+            onAuthorPress={() => {}}
+            onCommentGuidelinesPress={() => {}}
+            onCommentsPress={() => {}}
+            onLinkPress={() => {}}
+            onRelatedArticlePress={() => {}}
+            onTopicPress={() => {}}
+            onTwitterLinkPress={() => {}}
+            onVideoPress={() => {}}
+          />
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    },
+    {
+      name: "an inline image",
+      test() {
+        const testInstance = TestRenderer.create(
+          <Article
+            {...articleProps}
+            adConfig={adConfig}
+            analyticsStream={() => {}}
+            article={articleFixture({
+              ...testFixture,
+              ...emptyArticle,
+              content: [
+                {
+                  attributes: {
+                    caption: "A Caption",
+                    credits: "Some Credits",
+                    display: "inline",
+                    ratio: "9:4",
+                    url: "https://image-inline.io"
+                  },
+                  children: [],
+                  name: "image"
+                }
+              ]
+            })}
+            onAuthorPress={() => {}}
+            onCommentGuidelinesPress={() => {}}
+            onCommentsPress={() => {}}
+            onLinkPress={() => {}}
+            onRelatedArticlePress={() => {}}
+            onTopicPress={() => {}}
+            onTwitterLinkPress={() => {}}
+            onVideoPress={() => {}}
+          />
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    }
+  ];
+
+  iterator(tests);
 };
