+++ packages/article/__tests__/images.web.js
@@ -1,3 +1,5 @@
+import React from "react";
+import TestRenderer from "react-test-renderer";
 import {
   addSerializers,
   enzymeTreeSerializer,
@@ -6,8 +8,12 @@ import {
   minimalWebTransform,
   print
 } from "@times-components/jest-serializer";
+import { iterator } from "@times-components/test-utils";
+import Article from "../src/article";
+import articleFixture, { testFixture } from "../fixtures/full-article";
+import { adConfig } from "./ad-mock";
+import articleProps from "./shared-article-props";
 import "./mocks.web";
-import shared from "./images.base";
 
 const omitProps = new Set([
   "className",
@@ -15,6 +21,16 @@ const omitProps = new Set([
   "responsiveLinkStyles",
   "style"
 ]);
+
+const emptyArticle = {
+  byline: null,
+  flags: null,
+  label: null,
+  leadAsset: null,
+  relatedArticleSlice: null,
+  standfirst: null,
+  topics: null
+};
 
 export default () => {
   addSerializers(
@@ -27,5 +43,86 @@ export default () => {
     )
   );
 
-  shared();
+  const tests = [
+    {
+      name: "a secondary image",
+      test() {
+        const testInstance = TestRenderer.create(
+          <Article
+            {...articleProps}
+            adConfig={adConfig}
+            analyticsStream={() => {}}
+            article={articleFixture({
+              ...testFixture,
+              ...emptyArticle,
+              seoContent: [
+                {
+                  attributes: {
+                    caption: "A Caption",
+                    credits: "Some Credits",
+                    display: "secondary",
+                    ratio: "3:2",
+                    url: "https://image-2.io"
+                  },
+                  children: [],
+                  name: "image"
+                }
+              ]
+            })}
+            onAuthorPress={() => {}}
+            onCommentGuidelinesPress={() => {}}
+            onCommentsPress={() => {}}
+            onLinkPress={() => {}}
+            onRelatedArticlePress={() => {}}
+            onTopicPress={() => {}}
+            onTwitterLinkPress={() => {}}
+            onVideoPress={() => {}}
+          />
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    },
+    {
+      name: "an inline image",
+      test() {
+        const testInstance = TestRenderer.create(
+          <Article
+            {...articleProps}
+            adConfig={adConfig}
+            analyticsStream={() => {}}
+            article={articleFixture({
+              ...testFixture,
+              ...emptyArticle,
+              seoContent: [
+                {
+                  attributes: {
+                    caption: "A Caption",
+                    credits: "Some Credits",
+                    display: "inline",
+                    ratio: "9:4",
+                    url: "https://image-inline.io"
+                  },
+                  children: [],
+                  name: "image"
+                }
+              ]
+            })}
+            onAuthorPress={() => {}}
+            onCommentGuidelinesPress={() => {}}
+            onCommentsPress={() => {}}
+            onLinkPress={() => {}}
+            onRelatedArticlePress={() => {}}
+            onTopicPress={() => {}}
+            onTwitterLinkPress={() => {}}
+            onVideoPress={() => {}}
+          />
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    }
+  ];
+
+  iterator(tests);
 };
