+++ packages/article/__tests__/web/article-lazy.web.test.js
@@ -92,7 +92,7 @@ iterator([
     test() {
       const article = articleFixture({
         ...testFixture,
-        content: [
+        seoContent: [
           {
             attributes: {
               caption: "An image caption",
@@ -138,7 +138,7 @@ iterator([
     test() {
       const article = articleFixture({
         ...testFixture,
-        content: [
+        seoContent: [
           {
             attributes: {
               caption: "An image caption",
