+++ packages/article/fixtures/full-article.js
@@ -1982,6 +1982,7 @@ const makeDefaultConfig = ({
   leadAsset = defaultLeadAsset,
   relatedArticleSlice = defaultRelatedArticleSlice,
   section = defaultSection,
+  seoContent = defaultContent,
   shortHeadline = defaultShortHeadline,
   shortIdentifier = defaultShortIdentifier,
   slug = defaultSlug,
@@ -2001,6 +2002,7 @@ const makeDefaultConfig = ({
   leadAsset,
   relatedArticleSlice,
   section,
+  seoContent,
   shortHeadline,
   shortIdentifier,
   slug,
