+++ packages/provider-test-tools/fixtures/article.json
@@ -16,7 +16,7 @@
   ],
   "commentCount": 65,
   "commentsEnabled": true,
-  "content": [
+  "seoContent": [
     {
       "name": "paragraph",
       "attributes": {},
